requirements.txt
transformers==4.38.2
accelerate==1.10.1
bitsandbytes==0.41.3
peft==0.15.2
einops==0.8.1
huggingface_hub==0.31.2
tokenizers==0.21.1
sentencepiece # Required for the LLaMA tokenizer

opencv-python>=4.6
scipy==1.11.4

numpy==1.26.4

Pillow==11.2.1
Requests==2.32.3
tqdm==4.67.1
matplotlib

decord==0.6.0
deepspeed==0.16.7
einops_exts==0.0.4
packaging==25.0
triton==2.1.0